+++ scout/server/app.py
@@ -157,9 +157,9 @@ def configure_extensions(app):
         # setup connection to google oauth2
         configure_oauth_login(app)
 
-    if app.config.get("CLOUD_IGV_TRACKS"):
-        LOG.info("Collecting IGV tracks from cloud resources")
-        extensions.cloud_tracks.init_app(app)
+    if app.config.get("CUSTOM_IGV_TRACKS") or app.config.get("CLOUD_IGV_TRACKS"):
+        LOG.info("Collecting IGV tracks from cloud or local resources")
+        extensions.config_igv_tracks.init_app(app)
 
     if app.config.get("PHENOPACKET_API_URL"):
         LOG.info("Enable Phenopacket API")